Crypto/requirements.txt
pandas
numpy
python-binance
matplotlib
seaborn
optuna